Move mapping annotations that can be used outside mapping context.
