+++ implementation/src/main/java/io/smallrye/config/WithDefault.java
@@ -1,4 +1,4 @@
-package io.smallrye.config.mapper;
+package io.smallrye.config;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -7,16 +7,11 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * The name of the configuration property or group.
+ * Specify the default value of a property.
  */
 @Documented
-@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface WithName {
-    /**
-     * The name of the property or group. Must not be empty.
-     *
-     * @return the name
-     */
+@Target(ElementType.METHOD)
+public @interface WithDefault {
     String value();
 }
