+++ implementation/src/main/java/io/smallrye/config/WithName.java
@@ -1,4 +1,4 @@
-package io.smallrye.config.mapper;
+package io.smallrye.config;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -7,11 +7,16 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Specify the default value of a property.
+ * The name of the configuration property or group.
  */
 @Documented
-@Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface WithDefault {
+@Retention(RetentionPolicy.RUNTIME)
+public @interface WithName {
+    /**
+     * The name of the property or group. Must not be empty.
+     *
+     * @return the name
+     */
     String value();
 }
